Comment out the new rules
The rule additions aren't a part of the build system improvement but
they are an improvement of the profile. The change in profile should
be submitted as a separate PR to separate content discussions from
build system discussions. However, the profile is an incomplete draft,
so extraction may not be worth the effort. Therefore we are only
changing them to comments instead of extracting them.